added working food activity with a menu!
- made a food menu that activates via the activity system, showing a page that has several food buttons. buttons activate a confirmation box when clicked, where if yes is selected, the corresponding food is given.
- in programming terms, each button has a script with a ref to a corresponding food scriptableobj. the obj contains data for the sprite, and how much it should impact the hungry, happy, etc meters.
the food activity sets the buttons up to call its "open the confirmation box" function, which takes the scriptableobj from the clicked button's script, and disables the other buttons to avoid accidental clicking.
if yes is clicked, the food activity's "feed" function is called, and the confirm box is closed. at that point the menu would ideally temporarily disappear to play the animation of eating, but that's for when i start making the animation
- obviously still need to add the actual foods, but i'll do that once i've drawn the food assets as well